+++ core/basic_models/actions/rtdm_send_response_to_pp_action.py
@@ -10,7 +10,7 @@ from scenarios.user.user_model import User
 RTDM_RESPONSE = "RTDM_RESPONSE"
 
 
-class RtdmSendResponseToPoAction(CommandAction):
+class RtdmSendResponseToPpAction(CommandAction):
     """
     Action отправки отклика на персональное предложение (ПП) в Real-Time Decision Manager (RTDM). Аналитика:
     https://confluence.sberbank.ru/pages/viewpage.action?pageId=5786345296
