+++ smart_kit/resources/__init__.py
@@ -12,8 +12,7 @@ from core.basic_models.actions.counter_actions import CounterIncrementAction, Co
     CounterClearAction, CounterSetAction, CounterCopyAction
 from core.basic_models.actions.external_actions import ExternalAction
 from core.basic_models.actions.external_actions import ExternalActions
-from core.basic_models.actions.rtdm_get_pp_and_events_action import RtdmGetPpAndEventsAction
-from core.basic_models.actions.rtdm_send_response_to_po_action import RtdmSendResponseToPoAction
+from core.basic_models.actions.rtdm_send_response_to_pp_action import RtdmSendResponseToPpAction
 from core.basic_models.actions.push_action import PushAction
 from core.basic_models.actions.string_actions import StringAction, AfinaAnswerAction, SDKAnswer, \
     SDKAnswerToUser
@@ -205,6 +204,7 @@ class SmartAppResources(BaseConfig):
         ffd.field_filler_description["requirement"] = cffd.RequirementFiller
         ffd.field_filler_description["user_id"] = ffd.UserIdFiller
         ffd.field_filler_description["date_period_filler"] = ffd.DatePeriodFiller
+        ffd.field_filler_description["rtdm_get"] = ffd.RtdmGetPpAndEventsFiller
 
     def init_scenarios(self):
         scenarios[None] = BaseScenario
@@ -307,8 +307,7 @@ class SmartAppResources(BaseConfig):
         actions["set_local_variable"] = SetLocalVariableAction
         actions["set_variable"] = SetVariableAction
         actions["string"] = StringAction
-        actions["rtdm_get"] = RtdmGetPpAndEventsAction
-        actions["rtdm_send"] = RtdmSendResponseToPoAction
+        actions["rtdm_send"] = RtdmSendResponseToPpAction
         actions["push"] = PushAction
 
     def init_requirements(self):
